moved functions to app
