+++ src/App.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import "bootstrap/dist/css/bootstrap.min.css";
 import "./App.css";
 //import MovieItem from "./components/MovieItem";
@@ -6,7 +6,7 @@ import "./App.css";
 //import SearchInput from "./components/SearchInput";
 //import AddFavourites from "./components/AddFavourites";
 //import RemoveFavourites from "./components/RemoveFavourite";
-//import API_KEY from "./API_KEY";
+import API_KEY from "./API_KEY";
 import Sidebar from "./components/Sidebar/Sidebar";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import MyFavourites from "./components/MyFavourites/MyFavourites";
@@ -14,7 +14,41 @@ import Genres from "./components/Genres/Genres";
 import Search from "./components/Search/Search";
 
 const App = () => {
-  const [favourites] = useState([]);
+  const [searchInputValue, setSearchInputValue] = useState("hello");
+  const [movies, setMovies] = useState([]);
+  const [favourites, setFavourites] = useState([]);
+
+  //search functionality with variable input
+  const getMovieDetails = async (searchInputValue) => {
+    const url = `http://www.omdbapi.com/?s=${searchInputValue}&apikey=${API_KEY}`;
+
+    const response = await fetch(url);
+    const responseJSON = await response.json();
+
+    if (responseJSON.Search) {
+      setMovies(responseJSON.Search);
+    }
+  };
+
+  const addFavouriteMovie = (movie) => {
+    const newFavouriteList = [...favourites, movie];
+    setFavourites(newFavouriteList);
+    persistenceStorage(newFavouriteList);
+  };
+
+  const persistenceStorage = (items) => {
+    localStorage.setItem("movie-buff-faves", JSON.stringify(items));
+  };
+
+  useEffect(() => {
+    getMovieDetails(searchInputValue);
+  }, [searchInputValue]);
+  useEffect(() => {
+    const storedMovieFaves = JSON.parse(
+      localStorage.getItem("movie-buff-faves")
+    );
+    setFavourites(storedMovieFaves || []);
+  }, []);
 
   return (
     <Router>
@@ -25,11 +59,37 @@ const App = () => {
 
         <div className="main">
           <Routes>
-            <Route path="/search" exact element={<Search />} />
+            <Route
+              path="/search"
+              exact
+              element={
+                <Search
+                  searchInputValue={searchInputValue}
+                  setSearchInputValue={setSearchInputValue}
+                  movies={movies}
+                  setMovies={setMovies}
+                  favourites={favourites}
+                  setFavourites={setFavourites}
+                  getMovieDetails={getMovieDetails}
+                  addFavouriteMovie={addFavouriteMovie}
+                />
+              }
+            />
             <Route path="/genres" exact element={<Genres />} />
             <Route
               path="/myfavourites"
-              element={<MyFavourites favourites={favourites} />}
+              element={
+                <MyFavourites
+                  searchInputValue={searchInputValue}
+                  setSearchInputValue={setSearchInputValue}
+                  movies={movies}
+                  setMovies={setMovies}
+                  favourites={favourites}
+                  setFavourites={setFavourites}
+                  getMovieDetails={getMovieDetails}
+                  addFavouriteMovie={addFavouriteMovie}
+                />
+              }
             />
           </Routes>
         </div>
